Day 6

+++ day06/part2.c
@@ -1,4 +1,4 @@
-/* Day 6, part 2 = ? */
+/* Day 6, part 2 = 1675781200288 */
 
 #include <stdio.h>
 
@@ -6,19 +6,19 @@ int main() {
     FILE *inputFile = fopen("input.txt", "r");
 
     if (inputFile) {
-        long long timers[9] = {0, 0, 0, 0, 0, 0, 0, 0, 0};
-        long long timer;
+        long timers[9] = {0, 0, 0, 0, 0, 0, 0, 0, 0};
+        long timer;
         char ch;
-        long long answer = 0;
+        long answer = 0;
 
-        while (fscanf(inputFile, "%lld%c", &timer, &ch) >= 1) {
+        while (fscanf(inputFile, "%ld%c", &timer, &ch) >= 1) {
             ++timers[timer];
         }
 
         fclose(inputFile);
         
         for (int i = 0; i < 256; i++) {
-            long long newFishCount = *timers;
+            long newFishCount = *timers;
 
             for (timer = 0; timer < 8; timer++) {
                 timers[timer] = timers[timer + 1];
@@ -32,7 +32,7 @@ int main() {
             answer += timers[i];
         }
 
-        printf("%lld", answer);
+        printf("%ld", answer);
     }
 
     return 0;
